+++ app-code/chuck-norris-wrapper-service/src/test/java/com/example/chuck/norris/wrapper/integration/api/controller/ChuckNorrisControllerIntegrationTest.java
@@ -55,6 +55,5 @@ public class ChuckNorrisControllerIntegrationTest {
                 .isOk()
                 .expectBody()
                 .consumeWith(document("random-string", preprocessResponse(prettyPrint()) ));
-
     }
 }